DRTII-1381 Tweaks to hours breakdown results
